Fix E0308 to compile on apple hardware
`S_IWUSR` is of type `mode_t`:
https://github.com/rust-lang/libc/blob/d80e8bd3056d82fec80c022819084bdb0be3bdaa/src/unix/bsd/apple/mod.rs#L2928

On apple, `mode_t` is u16
https://github.com/rust-lang/libc/blob/d80e8bd3056d82fec80c022819084bdb0be3bdaa/src/unix/bsd/apple/mod.rs#L11

On linux, `mode_t` is u32
https://github.com/rust-lang/libc/blob/d80e8bd3056d82fec80c022819084bdb0be3bdaa/src/unix/linux_like/linux/mod.rs#L6

+++ src/cap.rs
@@ -32,7 +32,7 @@ use tokio::runtime::Handle;
 #[cfg(unix)]
 fn read_only(mut permissions: Permissions) -> Permissions {
   // Remove user write permissions.
-  let () = permissions.set_mode(permissions.mode() & !S_IWUSR);
+  let () = permissions.set_mode(permissions.mode() & !S_IWUSR as u32);
   permissions
 }
 
@@ -47,7 +47,7 @@ fn read_only(mut permissions: Permissions) -> Permissions {
 #[cfg(unix)]
 fn writeable(mut permissions: Permissions) -> Permissions {
   // Set user write permissions.
-  let () = permissions.set_mode(permissions.mode() | S_IWUSR);
+  let () = permissions.set_mode(permissions.mode() | S_IWUSR as u32);
   permissions
 }
